Refactor navn: 'containers' til 'sider'. Refactor navn: 'components' til 'felles-komponenter'
