+++ src/sider/arbeidsforholdDetalj.less
